requirements.txt
aiogram==2.20
aiohttp==3.8.1
aioice==0.7.6
aiortc==1.3.2
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
av==9.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
django-matrix-field==0.2
django-trench==0.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dlib==19.24.0
dnspython==2.2.1
frozenlist==1.3.0
google-crc32c==1.3.0
idna==3.3
imageio==2.19.3
install==1.3.5
kombu==5.2.4
multidict==6.0.2
netifaces==0.11.0
networkx==2.8.3
numpy==1.22.4
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
pyee==9.0.4
PyJWT==2.3.0
pylibsrtp==0.7.1
pyotp==2.6.0
pyparsing==3.0.9
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
redis==4.3.1
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
smsapi-client==2.6.0
sqlparse==0.4.2
tifffile==2022.5.4
twilio==7.8.2
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.7.2
yubico-client==1.13.0
